Cleanup and prep for V2
